requirements/test.txt
-r common.txt

coverage==7.8.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
pytest-env==1.1.5
pytest-sugar==1.0.0
pytest-order==1.3.0
ruff==0.11.5